Refactor ClusterProvider to remove legacy web3js dependency (#267)
This provider makes 3 requests, and makes their responses available to
child components:

- `getFirstAvailableBlock`
- `getEpochInfo`
- `getEpochSchedule`

The first 2 are reasonably straightforward, we have basically the same
structure as the legacy web3js, with just number/bigint changes.

`getEpochSchedule` is a bit more complex. The existing API exposes a
class with a bunch of functionality for finding the epoch for a slot,
and the first/last slot for an epoch. None of this is RPC functionality,
it's all baked into the legacy web3js code. Since the experimental
web3js doesn't do any of that, I've copied these functions into the
Explorer codebase, as pure functions that take an `EpochSchedule` (pure
data returned by the new RPC method) and a slot/epoch (bigint).

See existing web3js code here:
https://github.com/solana-labs/solana-web3.js/blob/9232d2b1019dc50f852ad70aa81624e751d76161/packages/library-legacy/src/epoch-schedule.ts

Also note that I hit a bug in experimental web3js where some of these
functions are incorrectly typed as unknown:
https://github.com/solana-labs/solana-web3.js/issues/1389
This was easy enough to work around for now

I've also moved `localStorageIsAvailable` from `utils/index.ts` to its
own `utils/local-storage`. This lets us import it without pulling in the
web3js dependency in `utils/index.ts`

The result of this PR is that the `ClusterProvider` in the root layout
no longer pulls in the legacy web3js dependency.